+++ test/utils/math/sequences/MathSequenceTest.kt
@@ -4,13 +4,13 @@ import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
 
-class DummyMathSequence : MathSequence(firstElement = 0L) {
-    override fun doGetNthElement(n: Int): Long {
-        return 0L
+class DummyMathSequence : MathSequence(firstElement = 0.0) {
+    override fun doGetNthElement(n: Int): Double {
+        return 0.0
     }
 
-    override fun doGetNFirstElementsSum(n: Int): Long {
-        return 0L
+    override fun doGetNFirstElementsSum(n: Int): Double {
+        return 0.0
     }
 }
